# Faster block copy functions.