… src/x86/sse2mc.c
… src/x86/sse2util.c
… src/x86/x86int.h
… src/x86/x86state.c
… src/x86/x86util.c