This improves block copying performance between `20%` and `50%` when SSE2 is available. 8x8, 4x4 and 2x2 did not show any improvement. 

```
16x16
sse2 0.518970s
   c 0.766928s
32x32
sse2 2.621024s
   c 3.184177s
64x64
sse2 5.729582s
   c 10.430384s
```