@mbebenita 2015-10-12
Reviewed @ https://review.xiph.org/993/